move arrayBufferToWordArray to libs
