+++ src/XAPI/XAPI.spec.ts
@@ -1,5 +1,6 @@
 import { XAPI, Agent, Statement, Activity, Attachment, AttachmentUsage } from ".";
 import * as CryptoJS from "crypto-js";
+import { arrayBufferToWordArray } from "../lib/arrayBufferToWordArray";
 import { TextEncoder } from "util";
 
 const endpoint: string = process.env.LRS_ENDPOINT || "";
@@ -29,17 +30,6 @@ const testStatement: Statement = {
     },
     object: testActivity
 };
-
-
-
-function arrayBufferToWordArray(ab: ArrayBuffer): any {
-    const i8a = new Uint8Array(ab);
-    const a = [];
-    for (let i = 0; i < i8a.length; i += 4) {
-        a.push(i8a[i] << 24 | i8a[i + 1] << 16 | i8a[i + 2] << 8 | i8a[i + 3]);
-    }
-    return CryptoJS.lib.WordArray.create(a, i8a.length);
-}
 
 describe("statement api", () => {
     test("can create a statement", () => {
