+++ src/lib/arrayBufferToWordArray.ts
@@ -1,0 +1,10 @@
+import * as CryptoJS from "crypto-js";
+
+export function arrayBufferToWordArray(ab: ArrayBuffer): CryptoJS.WordArray {
+  const i8a = new Uint8Array(ab);
+  const a = [];
+  for (let i = 0; i < i8a.length; i += 4) {
+      a.push(i8a[i] << 24 | i8a[i + 1] << 16 | i8a[i + 2] << 8 | i8a[i + 3]);
+  }
+  return CryptoJS.lib.WordArray.create(a, i8a.length);
+}
